LayoutTests/imported/w3c/web-platform-tests/xhr/setrequestheader-case-insensitive-expected.txt


PASS XMLHttpRequest: setRequestHeader() - headers that differ in case
FAIL XMLHttpRequest: setRequestHeader() - headers that differ in case 1 assert_array_equals: expected property 0 to be "content-TYPE" but got "Content-Type" (expected array ["content-TYPE"] got ["Content-Type"])
